requirements.txt
cffi==1.15.0
future==0.18.2
paho-mqtt==1.6.1
pycparser==2.21
python-olm==3.1.3
